+++ R/io.R
@@ -2,6 +2,8 @@
 #'
 #' @param fname Name of target FHX file. Needs to be in format version 2.
 #' @param encoding Encoding to use when reading the FHX file. The default is to use the system.
+#' @param text Character string. If \code{fname} is not provided and 
+#'   \code{text} is, then data is read from \code{text} using a text connection.
 #'
 #' @return An \code{fhx} object.
 #'
@@ -11,16 +13,23 @@
 #' }
 #'
 #' @export
-read_fhx <- function(fname, encoding=getOption("encoding")) {
-  con <- file(fname, encoding = encoding)
-  on.exit(close(con))
+read_fhx <- function(fname, encoding, text) {
+  if (missing(encoding))
+    encoding <- getOption('encoding')
+  if (missing(fname) && !missing(text)) {
+    con <- textConnection(text)
+    on.exit(close(con))
+  }
+  if (!missing(fname)) {
+    con <- file(fname, encoding = encoding)
+    on.exit(close(con))
+  }
   # Error checking and basic variables.
   if (length(readLines(con, n = 1)) == 0)
     stop("file appears to be empty")
   fl <- readLines(con, warn = FALSE)
   if (!any(suppressWarnings(grepl("FHX2 FORMAT|FIRE2 FORMAT", fl, ignore.case = TRUE))))
     stop("Cannot find line 'FHX2 FORMAT' or 'FIRE2 FORMAT'.")
-
   first <- suppressWarnings(grep("FHX2 FORMAT|FIRE2 FORMAT", fl, ignore.case = TRUE))
   describe <- as.numeric(strsplit(fl[[first + 1]], " ")[[1]])
   if (length(describe) != 3) {  # First year; no. sites; length of site id.